// src/Enums/Enums.ts
export enum Windows {
    START = "start",
    AUTH = "authorization",
    REG = "registration",
    LOGIN = "log in",
    FARMS = "farms",
    GAME = "game"
}

export enum GameFieldWindows {
    GAME = "game",
    STATISTICS = "statistics",
    FARM_REVIEW = "farm review",
    PET_REVIEW = "pet review",
    FRIENDS = "friends",
    SETTINGS = "settings",
    CONSTRUCTOR = "constructor"
}

export enum BodyTypesEnum {
    PEAR = "pear",
    OVAL = "oval",
    EGG = "egg",
    SQUARE = "square",
    CIRCLE = "circle"
}

export enum BodyPartTypesEnum {
    BODIES = "bodies",
    MOUTHS = "mouths",
    NOSES = "noses",
    EYES = "eyes"
}

export enum BodyPartSizesEnum {
    BODIES = 7,
    MOUTHS = 4,
    NOSES = 3,
    EYES = 4
}